created the model of the motor casing
I tried to specify the mass and CoM. Fudging the density and adding an
internal cavity would be a way to do it, if we really need to. The
"official" way to do it is with a design table. However, that feature on
SW is actually... an Excel integration. So, nope.